manager/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docker==4.2.1
fastapi==0.55.1
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
pydantic==1.5.1
pymongo==3.10.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
starlette==0.13.2
urllib3==1.25.9
websocket-client==0.57.0
